[breaking] hbotapi structuration
