+++ honeybot/plugins/greet.py
@@ -20,6 +20,7 @@ class Plugin:
         pass
 
     def run(self, incoming, methods, info, bot_info):
+        
         try:
             # if '!~' in info['prefix']:
                 # print(info)
