--- honeybot/plugins_info.md
@@ -1,341 +1,0 @@
-# abbreviation.py
-by Erin Moon, Sourav Dutta\
-responds to .def lol
-\
-<code>.def lol</code>
-
-# age.py
-by Justin Walker\
-Given birthday it returns your age.
-\
-<code>.age \<\<day>> \<\<month>> \<\<year>></code>
-
-# bitcoin.py
-by Gabriele Ron, [Website](https://Macr0Nerd.github.io)\
-Checks the current price for Bitcoin through the Legacy Coin Market Cap API
-\
-<code>.btc</code>
-
-# blackjack.py
-by Angelo Giacco\
-Play blackjack
-\
-<code>.blackjack create</code>
-
-# book_of_day.py
-by Schoberg, UMD\
-using bookoftheday.org, get the title and link of the book of the day
-\
-<code>.bookoftheday</code>
-
-# caesar_cipher.py
-by Kyle Galloway\
-encrypts and decrypts using a caesar cipher
-\
-<code>.caesar_cipher \<message></code>
-
-# calc.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-evaluates maths expressions in the format supported by py
-\
-<code>.calc \<maths expression></code>
-
-# channeljoin.py
-by Marcelo Benesciutti\
-Bot Will join a given channel in the server
-\
-<code>.channeljoin channel</code>
-
-# cointoss.py
-by Shriram Bhat\
-Responds to .cointoss, Returns head or tail at random
-\
-
-# comic.py
-by Miguel Boekhold\
-Returns a random comic from xkcd
-\
-<code>.comic</code>
-
-# converter.py
-by Angelo Giacco\
-Converts currencies
-\
-<code>.convert \<\<base currency code>> \<\<target currency code>> \<\<amount>></code>
-
-# conv_sniff.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-senses conversation topic with sensitivity set by user, also supports
-\
-
-# country.py
-by Phil Beld\
-Show the capital of a given country.
-\
-<code>.country \<country></code>
-
-# dadjoke.py
-by moffetma Oregon State University CS\
-Sends a random "dad joke" from icanhazdadjoke.com
-\
-<code>.dadjoke</code>
-
-# date.py
-by Gico Carlo Evangelista, https://gicocarlo.me/\
-Will post the current date
-\
-<code>.date today</code>
-
-# debug.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-prints all parameters passed to bot
-\
-<code>.debug</code>
-
-# dictionary.py
-by Nishant, JPMorgan Chase & Co.\
-sends the meaning of the word
-\
-<code>.dictionary \<\<word>></code>
-
-# fact.py
-by Justin Walker\
-Returns a random fact.
-\
-<code>.fact</code>
-
-# google.py
-by Justin Walker\
-Returns the first three links from a google search.
-\
-<code>.google \<\<search term>></code>
-
-# greet.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-responds to .hi, demo of a basic plugin
-\
-<code>.hi</code>
-
-# hangman.py
-by Justin Walker\
-Plays game of hangman
-\
-<code>.hangman start</code>
-
-# help.py
-by Eduardo Moraes de Mello, https://github.com/edumello\
-Show link with all the functionalities of the bot.
-\
-<code>.help</code>
-
-# horoscope.py
-by Angelo Giacco\
-Returns your daily horoscope based on your star sign
-\
-<code>.horoscope \<\<starsign>></code>
-
-# installed_modules.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-checks if all listed in requirements.txt installed
-\
-<code>.installed</code>
-
-# joins.py
-by Gico Carlo Evangelista, https://gicocarlo.me/\
-Greets everyone who joins the channel
-\
-
-# joke.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-sends a random joke on request
-\
-<code>.joke</code>
-
-# log.py
-by Gico Carlo Evangelista, https://gicocarlo.me/\
-Logs the chat into a log.txt file
-\
-
-# magic_ball.py
-by Zakaria Talhami\
-Fortune-telling for question using a magic 8 ball
-\
-<code>.8ball \<\<question>></code>
-
-# mail.py
-by Tanner Fry\
-Will allow the user to send an email while in the chat
-\
-<code>.mail \<To email address> .body \<Text for the body of your email> .subject \<Text of subject for email></code>
-
-# maths.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-some maths related commands
-\
-<code>.sin \<number></code>
-
-# memo.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-responds to .memo, demo of a basic memory plugin
-\
-<code>.memo add \<key> \<value></code>
-
-# monero.py
-by A. Eigenbrot\
-Lookup the price of Monero in any currency you want (including other cryptos). Defaults to USD\
-<code>.xmr \[\<currency>\]</code>
-
-# monopoly.py
-by Angelo Giacco\
-Play monopoly
-\
-<code>>>>.monopoly create</code>
-
-# news.py
-by Angelo Giacco\
-Gets the top 10 headlines around the world from bbc world news
-\
-<code>.news</code>
-
-# onthisday.py
-by Kenneth Gargan\
-Gets an interesting fact about today!
-\
-<code>.onthisday</code>
-
-# password_generator.py
-by Nishant, JPMorgan Chase & Co.\
-sends different type of passwords
-\
-<code>.passgen \<\<length>></code>
-
-# proverb.py
-by Sushant Kumar\
-sends a random proverb on request
-\
-<code>.proverb</code>
-
-# pydocs.py
-by Milad H,\
-Searches the python documentatioin(pydocs) for the related keyword(s)
-\
-<code>.pydocs \<keyword></code>
-
-# pynews.py
-by Sam Deans\
-Finds the top five news items on python.org and prints them to the IRC
-\
-<code>.pynews</code>
-
-# pypi.py
-by Donald Lieu\
-Searches for Python packages on pypi.org
-\
-<code>.pypi \<search term></code>
-
-# quote.py
-by German Corpaz\
-Sends a random quote on request
-\
-<code>.quote</code>
-
-# repostats.py
-by Shreyansh Sancheti\
-Gets the number of forks and stars for a given repo
-\
-<code>.repostats url</code>
-
-# riddle.py
-by Angelo Giacco\
-Returns a riddle or whether guess at the answer was correct or not.
-\
-<code>.riddle</code>
-
-# roll.py
-by Glenn Toms\
-Will roll dice in the form of 1d4, 10d6 and return total.
-\
-<code>.roll \<\<dice>></code>
-
-# roman_numeral.py
-by Nick Wiley\
-Returns the roman numeral equivalent of the number inputted
-\
-<code>.roman \<number></code>
-
-# russian_roulette.py
-by Angelo Giacco\
-In the original russian roulette, you have a one sixth chance of committing
-\
-<code>.russian_roulette</code>
-
-# selfTrivia.py
-by Angel Jimenez\
-Returns a private trivia question
-\
-<code>.selfTrivia</code>
-
-# send_message.py
-by Justin Walker\
-Sends a message to another channel.
-\
-<code>.send #channel .u riceabove .m how are you?</code>
-
-# test.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-tests run
-\
-
-# todo.py
-by Himanshu Ranjan, [Website](https://github.com/h-ranjan1110)\
-Todo list keeps all your pending task .
-\
-<code>>>>.todo add \<Your Task Here></code>
-
-# tpbquote.py
-by Divyanshu Mehta\
-Extracts quotes and send random quote from Trailer Park Boys
-\
-<code>>>>.quote \<count></code>
-
-# transfer-rumour.py
-by Angelo Giacco\
-Collects transfer rumours from BBC sport
-\
-<code>.transfer-rumour</code>
-
-# translate.py
-by Ahmed Deeb\
-A Google Translate plugin
-\
-<code>. Google Translator object.</code>
-
-# uptime.py
-by Nick Wiley\
-Returns how long the bot has been running for
-\
-<code>.uptime</code>
-
-# username.py
-by Abdur-Rahmaan Janhangeer, pythonmembers.club\
-generates random usernames, used for inspiration
-\
-<code>.uname</code>
-
-# weather.py
-by Gabriele Ron\
-A plugin to get the weather of a location
-\
-<code>.weather \<city> \<country code></code>
-
-# wikipedia.py
-by Gabriele Ron, [Website](https://Macr0Nerd.github.io)\
-sends a wikipedia article on request based off of a search or random query
-\
-<code>.wiki \<command> \<topic></code>
-
-# corona.py
-by Angelo Giacco\
-Get coronavirus pandemic information
-\
-<code>.corona \<optional country></code>
