+++ honeybot/settings/PLUGINS.conf
@@ -1,60 +1,1 @@
-abbreviation
-age
-bitcoin
-blackjack
-book_of_day
-caesar_cipher
-calc
-channeljoin
-cointoss
-comic
-conv_sniff
-converter
-corona
-country
-dadjoke
-date
-debug
-diary
-dictionary
-fact
-flight
-google
 greet
-hangman
-help
-horoscope
-installed_modules
-joins
-joke
-log
-magic_ball
-maths
-memo
-monero
-monopoly
-news
-onthisday
-password_generator
-proverb
-pydocs
-pynews
-pypi
-quote
-random_color
-repostats
-riddle
-roll
-roman_numeral
-russian_roulette
-selfTrivia
-send_message
-test
-todo
-tpbquote
-transfer-rumour
-translate
-uptime
-username
-weather
-wikipedia
